+++ Graph Theory/Flow/MinCostFlow/potentialsPointer.cpp
@@ -18,7 +18,7 @@ struct Edge{
 struct Path{
 	Long node, weight;
 	Path(){}
-	Path(Long node,Long weight) : node(node) , weight(weight) {}
+	Path(Long node, Long weight) : node(node) , weight(weight) {}
 	bool operator >(const Path &P) const{
 		return weight > P.weight;
 	}
@@ -38,7 +38,7 @@ struct Graph{
 		}
 	}
 	
-	void addEdge(Long u, Long v, Long w, Long cost, bool dir) {
+	void addEdge(int u, int v, Long w, Long cost, bool dir) {
 		Edge *forward = new Edge(u , v , w, cost);
 		Edge *backward = new Edge(v , u , 0, -cost);
 		forward->rev = backward;
@@ -56,18 +56,18 @@ struct Graph{
 		}
 	}
 	
-	void spfa(Long s , Long n) { //O(E V)
-		for (Long i = 0; i < n; i++) pot[i] = INF;
-		queue<Long> q;
+	void spfa(int s , int n) { //O(E V)
+		for (int i = 0; i < n; i++) pot[i] = INF;
+		queue<int> q;
 		pot[s] = 0;
 		inQueue[s] = true;
 		q.push(s);
 		while (!q.empty()) {
-			Long u = q.front();
+			int u = q.front();
 			q.pop();
 			inQueue[u] = false;
 			for (Edge *e : adj[u]) {
-				Long v = e->to;
+				int v = e->to;
 				if (e->cap - e->flow > 0 && pot[u] + e->cost < pot[v]) {
 					pot[v] = pot[u] + e->cost;
 					if (!inQueue[v]) {
@@ -79,7 +79,7 @@ struct Graph{
 		}
 	}
 	
-	pair<Long, Long> dijkstra(Long s, Long t, Long n){ //O(E log V)
+	pair<Long, Long> dijkstra(int s, int t, int n){ //O(E log V)
 		//<flow, cost>
 		priority_queue<Path , vector<Path> , greater<Path>> q;
 		vector<Long> d(n, INF);
@@ -92,9 +92,9 @@ struct Graph{
 			Path p = q.top();
 			q.pop();
 			int u = p.node;
-			if(p.weight != d[u]) continue;
+			if (p.weight != d[u]) continue;
 			for (Edge *e : adj[u]) {
-				Long v = e->to;
+				int v = e->to;
 				Long cf = e->cap - e->flow;
 				Long cost = e->cost + pot[u] - pot[v];
 				if (cf > 0 && d[u] + cost < d[v]) {
@@ -106,10 +106,10 @@ struct Graph{
 			}
 		}
 		if (d[t] == INF) return {0, 0};
-		for (Long i = 0; i < n; i++) pot[i] += d[i];
+		for (int i = 0; i < n; i++) pot[i] += d[i];
 		Long cf = residualCap[t];
-		Long cur = t;
-		while(true ){
+		int cur = t;
+		while (true) {
 			parent[cur]->flow += cf;
 			parent[cur]->rev->flow -= cf;
 			cur = parent[cur]->from;
@@ -121,19 +121,19 @@ struct Graph{
 	}
 	
 	
-	pair<Long,Long> minCostFlow(Long s, Long t, Long n){ 
+	pair<Long, Long> minCostFlow(int s, int t, int n){ 
 		//O(E log V *  maxFlow)
 		//maxFlow <= V * U, where U is the maximum capacity
 		//Initially no negative cycles
 		//<maxFlow, minCost>
 		spfa(s , n); //not necessary if there is no negative edges
-		pair<Long,Long> inc;
-		pair<Long,Long> ans = {0, 0};
+		pair<Long, Long> inc;
+		pair<Long, Long> ans = {0, 0};
 		do {
 			inc = dijkstra(s , t , n );
 			ans.first += inc.first;
 			ans.second += inc.second;
-		} while(inc.first > 0);
+		} while (inc.first > 0);
 		return ans;
 	}
 } G;
